Created app and inout packages Removed redundant methods
