+++ pywps/app/basic.py
@@ -1,0 +1,26 @@
+import lxml
+from werkzeug.wrappers import Response
+from pywps import OWS, NAMESPACES
+
+
+def xpath_ns(el, path):
+    return el.xpath(path, namespaces=NAMESPACES)
+
+
+def xml_response(doc):
+    return Response(lxml.etree.tostring(doc, pretty_print=True),
+                    content_type='text/xml')
+
+
+class UOM(object):
+    """
+    :param uom: unit of measure
+    """
+
+    def __init__(self, uom=''):
+        self.uom = uom
+
+    def describe_xml(self):
+        return OWS.UOM(
+            self.uom
+        )
