+++ pywps/inout/__init__.py
@@ -1,0 +1,3 @@
+from pywps.inout.inputs import LiteralInput, ComplexInput
+from pywps.inout.outputs import LiteralOutput, ComplexOutput
+from pywps.inout.formats import Format, FORMATS
