+++ pywps/inout/storage.py
