--- pywps/outputs.py
@@ -1,246 +1,0 @@
-from pywps import inout
-from pywps import config
-from pywps._compat import text_type
-from pywps import E, WPS, OWS, XMLSCHEMA_2
-
-class LiteralOutput(inout.LiteralOutput):
-    """
-    :param identifier: The name of this output.
-    :param data_type: Type of literal input (e.g. `string`, `float`...).
-    :param value: Resulting value
-            Should be :class:`~String` object.
-    """
-
-    def __init__(self, identifier, title, data_type='string', abstract='', metadata=[], uom=[]):
-        inout.LiteralOutput.__init__(self, identifier, title=title, data_type=data_type)
-        self.abstract = abstract
-        self.metadata = metadata
-        self.uom = uom
-
-    def describe_xml(self):
-        doc = E.Output(
-            OWS.Identifier(self.identifier),
-            OWS.Title(self.title)
-        )
-
-        if self.abstract:
-            doc.append(OWS.Abstract(self.abstract))
-
-        for m in self.metadata:
-            doc.append(OWS.Metadata(m))
-
-        literal_data_doc = E.LiteralData()
-
-        if self.data_type:
-            literal_data_doc.append(OWS.DataType(self.data_type, reference=XMLSCHEMA_2 + self.data_type))
-
-        if self.uom:
-            default_uom_element = self.uom[0].describe_xml()
-            supported_uom_elements = [u.describe_xml() for u in self.uom]
-
-            literal_data_doc.append(
-                E.UOMs(
-                    E.Default(default_uom_element),
-                    E.Supported(*supported_uom_elements)
-                )
-            )
-
-        return doc
-
-    def execute_xml(self):
-        doc = WPS.Output(
-            OWS.Identifier(self.identifier),
-            OWS.Title(self.title)
-        )
-
-        if self.abstract:
-            doc.append(OWS.Abstract(self.abstract))
-
-        data_doc = WPS.Data()
-
-        literal_data_doc = WPS.LiteralData(text_type(self.data))
-        literal_data_doc.attrib['dataType'] = self.data_type
-        literal_data_doc.attrib['reference'] = XMLSCHEMA_2 + self.data_type
-        if self.uom:
-            default_uom_element = self.uom[0].describe_xml()
-            supported_uom_elements = [u.describe_xml() for u in self.uom]
-
-            literal_data_doc.append(
-                E.UOMs(
-                    E.Default(default_uom_element),
-                    E.Supported(*supported_uom_elements)
-                )
-            )
-        data_doc.append(literal_data_doc)
-
-        doc.append(data_doc)
-
-        return doc
-
-
-class ComplexOutput(inout.ComplexOutput):
-    """
-    :param identifier: The name of this output.
-    :param formats: Possible output formats for this output.
-            Should be list of :class:`~Format` object.
-    :param output_format: Required format for this output.
-            Should be :class:`~Format` object.
-    :param encoding: The encoding of this input or requested for this output
-            (e.g., UTF-8).
-    """
-
-    def __init__(self, identifier, title, formats=None, output_format=None, encoding="UTF-8",
-                 schema=None, abstract='', metadata=[]):
-        inout.ComplexOutput.__init__(self, identifier, title=title, abstract=abstract)
-        self.formats = formats
-        self.metadata = metadata
-
-        self._schema = None
-        self._output_format = None
-        self._encoding = None
-
-        self.as_reference = False
-        self.output_format = output_format
-        self.encoding = encoding
-        self.schema = schema
-        self.storage = None
-
-    @property
-    def output_format(self):
-        """Get output format
-        :rtype: String
-        """
-
-        if self._output_format:
-            return self._output_format
-        else:
-            return ''
-
-    @output_format.setter
-    def output_format(self, output_format):
-        """Set output format
-        """
-        self._output_format = output_format
-
-    @property
-    def encoding(self):
-        """Get output encoding
-        :rtype: String
-        """
-
-        if self._encoding:
-            return self._encoding
-        else:
-            return ''
-
-    @encoding.setter
-    def encoding(self, encoding):
-        """Set output encoding
-        """
-        self._encoding = encoding
-
-    @property
-    def schema(self):
-        """Get output schema
-        :rtype: String
-        """
-
-        return self._schema
-
-    @schema.setter
-    def schema(self, schema):
-        """Set output schema
-        """
-        self._schema = schema
-
-    def describe_xml(self):
-        default_format_el = self.formats[0].describe_xml()
-        supported_format_elements = [f.describe_xml() for f in self.formats]
-
-        doc = E.Output(
-            OWS.Identifier(self.identifier),
-            OWS.Title(self.title)
-        )
-
-        if self.abstract:
-            doc.append(OWS.Abstract(self.abstract))
-
-        for m in self.metadata:
-            doc.append(OWS.Metadata(*self.metadata))
-
-        doc.append(
-            E.ComplexOutput(
-                E.Default(default_format_el),
-                E.Supported(*supported_format_elements)
-            )
-        )
-
-        return doc
-
-    def execute_xml(self):
-        """Render Execute response XML node
-
-        :return: node
-        :rtype: ElementMaker
-        """
-
-        node = None
-        if self.as_reference:
-            node = self._execute_xml_reference()
-        else:
-            node = self._execute_xml_data()
-
-        doc = WPS.Output(
-            OWS.Identifier(self.identifier),
-            OWS.Title(self.title)
-        )
-        if self.abstract:
-            doc.append(OWS.Abstract(self.abstract))
-        doc.append(node)
-
-        return doc
-
-    def _execute_xml_reference(self):
-        """Return Reference node
-        """
-        doc = WPS.Reference()
-
-        # get_url will create the file and return the url for it
-        self.storage = FileStorage(config)
-        doc.attrib['{http://www.w3.org/1999/xlink}href'] = self.get_url()
-
-        if self.data_format:
-            if self.data_format.mime_type:
-                doc.attrib['mimeType'] = self.data_format.mime_type
-            if self.data_format.encoding:
-                doc.attrib['encoding'] = self.data_format.encoding
-            if self.data_format.schema:
-                doc.attrib['schema'] = self.data_format.schema
-        return doc
-
-    def _execute_xml_data(self):
-        """Return Data node
-        """
-        doc = WPS.Data()
-
-        if self.data is None:
-            complex_doc = WPS.ComplexData()
-        else:
-            complex_doc = WPS.ComplexData(self.data)
-
-        if self.data_format:
-            if self.data_format.mime_type:
-                complex_doc.attrib['mimeType'] = self.data_format.mime_type
-            if self.data_format.encoding:
-                complex_doc.attrib['encoding'] = self.data_format.encoding
-            if self.data_format.schema:
-                complex_doc.attrib['schema'] = self.data_format.schema
-        doc.append(complex_doc)
-        return doc
-
-
-class BoundingBoxOutput(object):
-    """bounding box output
-    """
-    # TODO: BoundingBoxOutput
-    pass
