+++ tests/test_literaltypes.py
@@ -1,7 +1,7 @@
 """Unit tests for IOs
 """
 import unittest
-from pywps.literaltypes import *
+from pywps.inout.literaltypes import *
 
 class ConvertorTest(unittest.TestCase):
     """IOHandler test cases"""
